Add common return clauses for gen_statem
gen_statem has a new set of return clauses in compare with gen_fsm.  One
example is the {keep_state, Data} which makes the state machine stay in
the current state.

Because there is two ways to implement a gen_statem, with either
state_functions or handle_event_function, we also need to take care of
the handle_event callback.